Basic JS functionality and page layout added
Added functions for each operation the calculator is capable of.
Added a function that takes the user input and the operand  and feeds it through to the appropriate function.
Added HTML and CSS to layout calculator on page dynamically so elements can be added or removed as the page is developed.